+++ client/src/main.jsx
@@ -1,4 +1,5 @@
 import { ThemeProvider } from '@material-tailwind/react';
+import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
